move dsv loader to file package
