+++ polars_hist_db/loaders/__init__.py
@@ -1,6 +1,6 @@
-from .dsv_loader import load_typed_dsv
-from .file_search import find_files
-from .ziptools import convert_zipped_csvs_to_parquet
+from .file.dsv_loader import load_typed_dsv
+from .file.file_search import find_files
+from .file.ziptools import convert_zipped_csvs_to_parquet
 
 __all__ = [
     "load_typed_dsv",
